Update to option 1
Restructured the way dependent information is recorded and processed.
Added a printout of employee information and option to confirm or start over
Restructured the way data is recorded to prevent recording of invalid information
Fixed dates showing with time 00:00;00 in data record
Added missing input for employee birth date
Fixed some minor details and mistakes


This option is pretty much finished now, moving on to other things. May add some extra validations to the inputs before its time to hand it in.