Added option 3: New inventory item
Added the function for "Option 3: New Inventory Item"

When this function is called, it will prompt the user for information about a new inventory item. The function will then confirm the information with the user, and save it into items.dat file.

Could add some additional validation to the inputs, but otherwise pretty easy.