Added option 4: Record Customer Purchase
This is a function that is included in the main program. The function is called when option 4 is selected. When its called, the program will prompt the user for information about one or more customer purchases.

Once the inputs are completed, the constant purchase_num is incremented by 1 and the information is written in a file called pruchases.dat .

This program also includes iteration over customers.dat to search if the customer is inside the data file. If they are not, the function will throw an error saying the customer is not found. For this to work, please always enter a name that matches any customer name in customers.dat

Once the inputs are complete, the information is summarized and the user is asked to confirm. Once confirmed, the information is written in purchases.dat

Still need to add input validation if there is time.